1. fixed signed/unsigned warnings: now functions return unsigned integers

+++ MigrationProject/user.c
@@ -30,12 +30,12 @@ void MRSG_set_task_cost_f ( double (*f)(enum phase_e phase, size_t tid, size_t w
     user.task_cost_f = f;
 }
 
-void MRSG_set_dfs_f ( void (*f)(char** dfs_matrix, size_t chunks, size_t workers, int replicas) )
+void MRSG_set_dfs_f ( void (*f)(char** dfs_matrix, size_t chunks, size_t workers, unsigned int replicas) )
 {
     user.dfs_f = f;
 }
 
-void MRSG_set_map_output_f ( int (*f)(size_t mid, size_t rid) )
+void MRSG_set_map_output_f ( size_t (*f)(size_t mid, size_t rid) )
 {
     user.map_output_f = f;
 }
